core: Move `render_children` into `TDisplayObjectContainer`.
