+++ core/src/display_object/button.rs
@@ -257,7 +257,7 @@ impl<'gc> TDisplayObject<'gc> for Button<'gc> {
     fn render(&self, context: &mut RenderContext<'_, 'gc>) {
         context.transform_stack.push(&*self.transform());
 
-        crate::display_object::render_children(context, self.iter_render_list());
+        self.render_children(context);
 
         context.transform_stack.pop();
     }
